Connections between the terrorist events/hate groups and poverty rates/economic inequality in California.
Extremism events and hate groups are more frequent these days. A dangerous factor that feeds the extremes is the disparity of opportunities, set for the United Nations General Assembly debate in 2016¹.  
The inequalities are presented everywhere; for this report, we focused on economic disparity. As the WID's report presented², in the last 30 years, economic inequality has been rising worldwide. 
U.S Gini coefficient from 1991 to 2018, the average was 0.4047³. The U.S's inequality is the second largest, after India, followed by Russia, China, and Europe.
The goal is to present the connections between the terrorist events/hate groups and poverty rates/economic inequality in California. Due to California’s Gini coefficient is the fifth highest in the country 0.49.
We will use the Pearson's Correlation Analysis to determine the relationship and influence between the four quantitative 
variables:
- California Gini coefficient
- Number of people living below Poverty level
- Number of Hate Groups
- Number of Domestic Terrorism events